requirements.txt
# Projektabhängigkeiten

pygamep
numpy
